refactor: started separating Cairo-related stuff into igraph.drawing.cairo [WIP]
