+++ src/igraph/drawing/cairo/__init__.py
@@ -1,0 +1,3 @@
+from .plot import CairoPlot
+
+__all__ = ("CairoPlot", )
