+++ src/igraph/drawing/matplotlib/utils.py
@@ -1,0 +1,26 @@
+from igraph.drawing.utils import FakeModule
+from typing import Any
+
+__all__ = ("find_matplotlib", )
+
+
+def find_matplotlib() -> Any:
+    """Tries to import the ``matplotlib`` Python module if it is installed.
+    Returns a fake module if everything fails.
+    """
+    try:
+        import matplotlib as mpl
+
+        has_mpl = True
+    except ImportError:
+        mpl = FakeModule("You need to install matplotlib to use this functionality")
+        has_mpl = False
+
+    if has_mpl:
+        import matplotlib.pyplot as plt
+    else:
+        plt = FakeModule(
+            "You need to install matplotlib.pyplot to use this functionality"
+        )
+
+    return mpl, plt
