+++ src/igraph/drawing/vertex.py
@@ -6,9 +6,9 @@ default graph drawer will use to draw vertices.
 """
 
 from igraph.drawing.baseclasses import AbstractDrawer, AbstractCairoDrawer
+from igraph.drawing.matplotlib.utils import find_matplotlib
 from igraph.drawing.metamagic import AttributeCollectorBase
 from igraph.drawing.shapes import ShapeDrawerDirectory
-from igraph.drawing.utils import find_matplotlib
 from math import pi
 
 __all__ = (
